finish rough draft of backend

+++ back_end/db/migrations/20190612021025_orders_table.js
@@ -12,7 +12,7 @@ exports.up = function(knex, Promise) {
       .defaultTo(-1)
 
     tbl
-      .integer('total')
+      .decimal('total')
       .notNullable()
 
     tbl
